requirements.txt
Django>=5.0
djangorestframework
drf-spectacular
pytest
pytest-django
